requirements.txt
# Core AWS
boto3>=1.34.0
botocore>=1.34.0

# Anthropic SDK (optional — for direct API calls outside Bedrock)
anthropic>=0.25.0

# Vector search (local FAISS for RAG embeddings)
faiss-cpu>=1.8.0
numpy>=1.26.0

# Neptune / Gremlin
gremlinpython>=3.7.0
requests>=2.31.0
requests-aws4auth>=1.3.0   # IAM signing for Neptune

# Data handling
pandas>=2.0.0
python-dotenv>=1.0.0

# Visualisation
matplotlib>=3.8.0
plotly>=5.20.0
kaleido>=0.2.1              # static PNG export for Plotly

# Token counting
tiktoken>=0.7.0

# Utilities
tqdm>=4.66.0
colorama>=0.4.6
tabulate>=0.9.0
